📦 Updated NuGet packages

+++ src/OpenRealEstate.Testing/ResidentialListingAssertHelpers.cs
@@ -27,7 +27,7 @@ namespace OpenRealEstate.Testing
             listing.Id.ShouldBe("Residential-Sold-ABCD1234");
             listing.StatusType.ShouldBe(StatusType.Sold);
 
-            decimal? soldPrice = 580000m;
+            int? soldPrice = 580000;
             listing.Pricing.SoldPrice.ShouldBe(soldPrice);
             listing.Pricing.SoldPriceText.ShouldBe(isSoldPriceVisibile
                                                        ? soldPrice.Value.ToString("C0")
